requirements.txt
gradio==3.24.1
openai==0.27.3
protobuf==4.22.1
pydub==0.25.1
python-dotenv==1.0.0
